+++ changelog.md
@@ -4,6 +4,7 @@
 
 * Added support for all integer, float, and char types.
 * Added `from_dict` to load Tuple from a vector in an "args" key.
+* Fixed a bug where keys did not preserve their order through an `include`.
 
 ## v1.3.0